Rename folder from `streaming` to `sources`
